Added VerboseAddressBalancesResult model
