+++ src/Stratis.Bitcoin.Features.BlockStore/AddressIndexing/AddressIndexRepository.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using LiteDB;
 using Microsoft.Extensions.Logging;
+using Stratis.Bitcoin.Controllers.Models;
 using Stratis.Bitcoin.Utilities;
 
 namespace Stratis.Bitcoin.Features.BlockStore.AddressIndexing
