+++ src/Stratis.Bitcoin/Controllers/Models/AddressBalancesResult.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using LiteDB;
 using NBitcoin;
 using Newtonsoft.Json;
 using Stratis.Bitcoin.Utilities.JsonConverters;
@@ -48,5 +49,56 @@ namespace Stratis.Bitcoin.Controllers.Models
         [JsonProperty("balance")]
         [JsonConverter(typeof(MoneyJsonConverter))]
         public Money Balance { get; set; }
+    }
+
+    /// <summary>
+    /// A class that contains a list of balances data per address as requested.
+    /// <para>Should the request fail the <see cref="Reason"/> will be populated.</para>
+    /// </summary>
+    public sealed class VerboseAddressBalancesResult
+    {
+        public VerboseAddressBalancesResult()
+        {
+            this.BalancesData = new List<AddressIndexerData>();
+        }
+
+        public List<AddressIndexerData> BalancesData { get; set; }
+
+        public int AddressIndexerHeight { get; set; }
+
+        public int ConsensusTipHeight { get; set; }
+
+        [JsonProperty("reason")]
+        public string Reason { get; private set; }
+
+        public static VerboseAddressBalancesResult RequestFailed(string reason)
+        {
+            return new VerboseAddressBalancesResult() { Reason = reason };
+        }
+    }
+
+    public class AddressIndexerData
+    {
+        [BsonId]
+        public string Address { get; set; }
+
+        public List<AddressBalanceChange> BalanceChanges { get; set; }
+    }
+
+    public class AddressBalanceChange
+    {
+        /// <summary><c>true</c> if there was a deposit to an address, <c>false</c> if it was a withdrawal.</summary>
+        public bool Deposited { get; set; }
+
+        public long Satoshi { get; set; }
+
+        /// <summary>Height of a block in which operation was confirmed.</summary>
+        public int BalanceChangedHeight { get; set; }
+
+        /// <inheritdoc />
+        public override string ToString()
+        {
+            return $"{nameof(this.Deposited)}:{this.Deposited}, {nameof(this.Satoshi)}:{this.Satoshi}, {nameof(this.BalanceChangedHeight)}:{this.BalanceChangedHeight}";
+        }
     }
 }
